refactor(expression): reorder expressions in build_from_prost, and use try_from pattern for InExpression and CaseExpression
