+++ src/expr/src/expr/mod.rs
@@ -46,9 +46,11 @@ use risingwave_pb::expr::ExprNode;
 
 use super::Result;
 use crate::expr::build_expr_from_prost::*;
+use crate::expr::expr_case::CaseExpression;
 use crate::expr::expr_coalesce::CoalesceExpression;
 use crate::expr::expr_concat_ws::ConcatWsExpression;
 use crate::expr::expr_field::FieldExpression;
+use crate::expr::expr_in::InExpression;
 use crate::expr::expr_nested_construct::NestedConstructExpression;
 use crate::expr::expr_regexp::RegexpMatchExpr;
 use crate::expr::expr_vnode::VnodeExpression;
@@ -94,6 +96,7 @@ pub fn build_from_prost(prost: &ExprNode) -> Result<BoxedExpression> {
     use risingwave_pb::expr::expr_node::Type::*;
 
     match prost.get_expr_type().unwrap() {
+        // Fixed number of arguments and based on `Unary/Binary/Ternary/...Expression`
         Cast | Upper | Lower | Md5 | Not | IsTrue | IsNotTrue | IsFalse | IsNotFalse | IsNull
         | IsNotNull | Neg | Ascii | Abs | Ceil | Floor | Round | BitwiseNot | CharLength
         | BoolOut | OctetLength | BitLength => build_unary_expr_prost(prost),
@@ -103,23 +106,27 @@ pub fn build_from_prost(prost: &ExprNode) -> Result<BoxedExpression> {
         | ConcatOp => build_binary_expr_prost(prost),
         And | Or | IsDistinctFrom | ArrayAccess => build_nullable_binary_expr_prost(prost),
         ToChar => build_to_char_expr(prost),
-        Coalesce => CoalesceExpression::try_from(prost).map(Expression::boxed),
-        Substr => build_substr_expr(prost),
         Length => build_length_expr(prost),
         Replace => build_replace_expr(prost),
-        Overlay => build_overlay_expr(prost),
         Like => build_like_expr(prost),
+        Repeat => build_repeat_expr(prost),
+        SplitPart => build_split_part_expr(prost),
+        Translate => build_translate_expr(prost),
+
+        // Variable number of arguments and based on `Unary/Binary/Ternary/...Expression`
+        Substr => build_substr_expr(prost),
+        Overlay => build_overlay_expr(prost),
         Trim => build_trim_expr(prost),
         Ltrim => build_ltrim_expr(prost),
         Rtrim => build_rtrim_expr(prost),
-        Repeat => build_repeat_expr(prost),
+
+        // Dedicated types
+        In => InExpression::try_from(prost).map(Expression::boxed),
+        Case => CaseExpression::try_from(prost).map(Expression::boxed),
+        Coalesce => CoalesceExpression::try_from(prost).map(Expression::boxed),
         ConcatWs => ConcatWsExpression::try_from(prost).map(Expression::boxed),
-        SplitPart => build_split_part_expr(prost),
         ConstantValue => LiteralExpression::try_from(prost).map(Expression::boxed),
         InputRef => InputRefExpression::try_from(prost).map(Expression::boxed),
-        Case => build_case_expr(prost),
-        Translate => build_translate_expr(prost),
-        In => build_in_expr(prost),
         Field => FieldExpression::try_from(prost).map(Expression::boxed),
         Array => NestedConstructExpression::try_from(prost).map(Expression::boxed),
         Row => NestedConstructExpression::try_from(prost).map(Expression::boxed),
